+++ parser_library/src/workspaces/library_local.h
@@ -16,6 +16,7 @@
 #define HLASMPLUGIN_PARSERLIBRARY_LOCAL_LIBRARY_H
 
 #include <atomic>
+#include <compare>
 #include <memory>
 #include <string>
 #include <string_view>
@@ -37,6 +38,26 @@ struct library_local_options
     std::vector<std::string> extensions;
     bool extensions_from_deprecated_source = false;
     bool optional_library = false;
+
+#ifdef __cpp_lib_three_way_comparison
+    auto operator<=>(const library_local_options&) const = default;
+#else
+    // libc++ (not even in main!!!)
+    bool operator<(const library_local_options& o) const
+    {
+        if (extensions < o.extensions)
+            return true;
+        if (extensions > o.extensions)
+            return false;
+        if (extensions_from_deprecated_source < o.extensions_from_deprecated_source)
+            return true;
+        if (extensions_from_deprecated_source > o.extensions_from_deprecated_source)
+            return false;
+        if (optional_library < o.optional_library)
+            return true;
+        return false;
+    }
+#endif
 };
 
 // library holds absolute path to a directory and finds macro files in it