+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 #### Changed
 - Macro label is the preferred go to definition target unless the request is made from the label itself
+- Library contents are now shared between processor groups
 
 ## [1.5.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.4.0...1.5.0) (2022-11-02)